Fixes all left issues with the fetch.
Reduces the concurrency to 1 for the character fetch which was causing
some data to never be obtained/saved.
Sets the items as array if empty, else it pushes the items in the
existing array.
Sets the date removed for ex-characters.
Sets the activities if loot or bosskill with bosskillID.
Sets the timeout at 120 seconds for requests.
Outputs the request that loaded to an errorhtml page.
Adds the riding profession.
Adds many logs for professions, reputations, classes, races, etc., that
might add null values in the database.